Add robust SQL and parameter compiling.
The previous framework for creating a SQL string ready for parameter
binding and execution involved a convoluted system that separated
paramaters from clauses. Now, parameters and clauses are treated as
nodes in a tree and the sub-trees are brought together upon compiling.
All of the parameters for a compiling clauses are extracted and unnamed
parameters are given a named parameter based on the number of current
parameters in the clause.

The 'value' function was eliminated from the Expr prototype to clean up
the syntax. Now, all binary operators accept the right operand as an
argument to the function.