Add argument for binary operator functions of Expr as optional.
To provide greater flexibility in creating a SQL expression, Expr, the
argument to functions representing binary operators is now optional.
Since the right operand is simply pushed onto the tree, it does not
matter if it is passed as an argument or added by function
chaining/cascading. This offers greater flexibility for users in
constructing expressions.