Refactor 'sql' to 'toString' for creating schema.
The 'sql' function was essentially renamed to the 'toString' function
and the old 'toString' function was removed. This was to reduce
redundant and unnecessary code. The original 'toString' function was
added to the prototypes of the 'Table', 'Column', and 'ForeignKey'
objects for debugging purposes, but the SQL is just as informative from
a debugging stand point. Additionally, the 'toString' function follows
the convention implemented with the 'Compiled' object, where the
'toString' function returns a SQL string ready for parameter binding and
execution.