Added factory methods for creating database schema elements.
The more conventional usage of a library is to avoid the use of the
'new' operator and reduce typing. Thus, factory methods, or wrapper
functions, were implemented as part of the library API instead of
exposing the constructors.

It appears from reviewing other JavaScript library and project code and
reading forum posts that constructors within a library should not be
called directly, but exposed through functions attached to a namespace
or module. Factory methods were created to follow this observation.