Fix named parameter key generation.
The automatic named parameter key generation for indexed parameters, or
parameters without explicit names, would replace the key for a Param
after a compile was conducted. This would result in a memory effect
where the named parameter assigned to the parameter during the compile
would be used as the new key for the parameter even after compiling. The
problem would be multiple named parameters with the same parameter name
if multiple compiles were conducted or if the parameter was used in more
than one statement.

A temporary key is created and used for the compiled statement, but the
key is not changed for the parameter. The key for an indexed parameter,
'?', should be the same before and after compiling using the 'toString'
function. This was not the case before the fix.