Add SQL string builder module.
Started moving SQL string building from Spengler module to separate
module. The Spengler module is a SQL expression language, and it is an
abstract from 'raw' SQL construction. However, the SQL string building
functionality was originally included in Spengler and this is very
low-level functionality without much abstraction. The SQL string
building functionality maybe desired without the overhead of the SQL
expression language or a different SQL string building module might be
used. The Spengler module still needs to implement a mechanism for using
alterative SQL string building modules, but this is in the future.

The Egon module will be built on top of the Spengler module which will
be built on top of the SQL string building module. For now, the SQL
string building module is simply named 'SQL'. A less generic and more
unqiue name will be needed in the future.