Implement individual clause creation
All clauses can be created independent of the chain and added either
through function chaining or in constructors. Almost every SQL string is
now a clause.

Type checking of arguments and optional arguments have been created. The
type checking currently does not throw an exception, as an
IllegalArgument exception still needs to be implemented.

Different types of join sources are now possible. Table, Source, and
Select join sources can be created. A Table source is created
automatically if a string is used for the join source.

Added additional documentation.