Refactor table insert to use Ramis library for statement creation
Added a function to the Ramis library to compile a clause into a
statement. A statement is a clause that has been 'frozen' into a SQL
string and bind parameters ready for binding and execution.

The execution of the statement is left to the script or library using
Ramis. In this case, it is the Spengler SQL expression language library.
This is continue decoupling and modularizing the codebase.

Signed-off-by: volks73 <cfield2@gmail.com>